moving `WasmValueLayout` constants from `raw/func_args` to `raw/wasm/layout`
